requirements.txt
Pillow
ReportLab
pillow_heif
pyinstaller
